% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{as.ez}
\alias{as.ez}
\title{Convert a data.frame or data.table object into an ez.data.frame.}
\usage{
as.ez(df, meta = NULL)
}
\arguments{
\item{meta}{data.frame containing meta information with at least two columns.}
}
\description{
Convert a data.frame or data.table object into an ez.data.frame.
}
\examples{
data("iris")
aa = as.ez(iris, meta = data.frame(names(iris), c('花萼长', '花萼宽', '花瓣长', '花瓣宽', '品种')))
class(aa)
attr(aa, 'meta')
table(aa, ~Species)
}
\seealso{
Other res: \code{\link{readSPSS}}, \code{\link{readStata}}
}

