% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ftable.ez.data.frame}
\alias{ftable.ez.data.frame}
\title{flat table using ftable()}
\usage{
\method{ftable}{ez.data.frame}(ez, formula, style = 1, prop_margin = 1, ...)
}
\arguments{
\item{style}{1=frequency, 2=percentage, 3=percentage with Sum}

\item{...}{parameters for ftable.formula(), must have both left and right hand sides.}
}
\description{
flat table using ftable()
}

