% Generated by roxygen2 (4.1.1.9001): do not edit by hand
% Please edit documentation in R/ez.R
\name{readStata}
\alias{readStata}
\title{Import Stata .dta file using package haven}
\usage{
readStata(file, encoding = NULL, charEncoding = encoding)
}
\arguments{
\item{file}{dta file name.}

\item{encoding}{Character encoding for labels, default is UTF-8.}

\item{charEncoding}{Character encoding for character variables, default is the same as encoding.}
}
\value{
ez.data.frame class object inherited from data.table
}
\description{
Import Stata .dta file using package haven
}
\examples{
library(ezdf)
dat = readStata('CGSS2013（居民问卷） 发布版.dta')
class(dat)
# set encoding properly
dat = readStata('CGSS2013（居民问卷） 发布版.dta', encoding = 'GB2312')
tbl(dat, a66 ~ s5a)
}
\seealso{
readSPSS

Other res: \code{\link{as.ez}}; \code{\link{readSPSS}}
}

