% Generated by roxygen2 (4.1.1.9001): do not edit by hand
% Please edit documentation in R/ez.R
\name{set_value_labels}
\alias{set_value_labels}
\title{assign value labels to an ez.data.frame object}
\usage{
set_value_labels(ez, col, labels)
}
\arguments{
\item{ez}{}

\item{col}{column name, character}

\item{labels}{a named vector for value labels, e.g. c(a=1, b=2, c=3)}
}
\description{
assign value labels to an ez.data.frame object
}
\examples{
library(ezdf)
data(iris)
iris = as.ez(iris)
class(iris)
iris$test = sample(5, size = nrow(iris), replace = T)
tbl(iris, ~test)
options('ezdfKeepVal' = T)
set_value_labels(iris, 'test', c(a=1, b=2, c=3))
tbl(iris, ~test, )
}

