% Generated by roxygen2 (4.1.1.9001): do not edit by hand
% Please edit documentation in R/ez.R
\name{tbl}
\alias{tbl}
\title{multivariate table}
\usage{
tbl(ez, expr, func = "mean", N = FALSE, sort = TRUE)
}
\arguments{
\item{expr}{formula object}

\item{func}{statistics applied on the data subset}

\item{DT}{ez.data.frame object}
}
\description{
multivariate table
}
\examples{
tbl(dat, a66 ~ s5a)
tbl(dat, a66 ~ s5a + s41, 'sum')
tbl(dat, a66 ~ s5a + s41, 'mean', N = T)
pander(tbl(dat, ~ s41 + s5a + a66, 'sum', sort = T))
tbl(dat, ~ s5a + a66)
}

