% Generated by roxygen2 (4.1.1.9001): do not edit by hand
% Please edit documentation in R/ez.R
\name{valueLabels<-}
\alias{valueLabels<-}
\title{assign value labels to an ez.data.frame object}
\usage{
valueLabels(ez, col) <- value
}
\arguments{
\item{ez}{}

\item{col}{column name, character}

\item{labels}{a named vector for value labels, e.g. c(a=1, b=2, c=3)}
}
\description{
assign value labels to an ez.data.frame object
}
\examples{
library(ezdf)
data(iris)
iris = as.ez(iris)
class(iris)
iris$test = sample(5, size = nrow(iris), replace = T)
tbl(iris, ~test)
options('ezdfKeepVal' = T)
valueLabels(iris, 'test') = c(a=1, b=2, c=3)
tbl(iris, ~test, )
}

