% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{varLabels}
\alias{varLabels}
\title{Get variable labels}
\usage{
varLabels(ez, varnames, default = "")
}
\arguments{
\item{default}{default value for when `varnames` exists or labels not 
set. Value of c("", "var"). Default is "". "var" for copying varnames.}
}
\description{
Get variable labels
}

